fix int stuff

+++ src/actor/projectile/HoningProjectile.java
@@ -18,7 +18,7 @@ public class HoningProjectile extends Projectile{
 		this.honing = true;
 	}
 	
-	public HoningProjectile(String name, double x_position, double y_position, Shape hit_box, Map map, 
+	public HoningProjectile(String name, int x_position, int y_position, Shape hit_box, Map map, 
 		double x_velocity, double y_velocity, double x_acceleration, double y_acceleration, boolean honing){
 		super(name, x_position, y_position, hit_box, 0, 1, map, x_velocity, y_velocity, x_acceleration, y_acceleration, true);
 		this.time_alive = 2;
